bigtable/requirements.txt
google-cloud-bigtable==0.30.0
google-cloud-core==0.28.1